Compiler warnings (#122)
* fix compiler warning C4244 "implicit conversion of uint32_t to float tuple"
(use static_cast to float to avoid implicit conversion)

* fix compiler warning C4244: conversion of double to float in return statement
(static_cast return value to function template parameter)

* fix compiler warnings C4267: conversion of size_t to uint32_t/unsigned int

* fix compiler warning C4244: conversion of double to float

* static cast size value to int to fix:
Warning: C4267 conversion of size_t to int

* static cast size value to int to fix:
Warning: C4267 conversion of size_t to int

* static cast size value to int to fix:
Warning: C4267 conversion of size_t to int

* use double instead of float to store get_time() return value

* static cast size value to uint32_t for Vulkan API

+++ examples/multi_invokee_rendering/source/multi_invokee_rendering.cpp
@@ -162,7 +162,7 @@ int main() // <== Starting point ==
 			if (values.size() > 90) {
 				values.erase(values.begin());
 			}
-			ImGui::PlotLines("ms/frame", values.data(), values.size(), 0, nullptr, 0.0f, FLT_MAX, ImVec2(0.0f, 100.0f));
+			ImGui::PlotLines("ms/frame", values.data(), static_cast<int>(values.size()), 0, nullptr, 0.0f, FLT_MAX, ImVec2(0.0f, 100.0f));
 			ImGui::End();
 		});
